finishes partial beet mover scheduling, adds tests
