+++ releasetasks/templates/l10n.yml.tmpl
@@ -352,7 +352,7 @@
 
 # repacks beetmover
 {% if push_to_candidates_enabled %}
-{% set partial_beetmover_buildername = "{}_partial_{}build{}_beetmover_candidates_{}".format(buildername, chunk, partial_version, partial_info["buildNumber"]) %}
+{% set partial_beetmover_buildername = "{}_partial_{}build{}_beetmover_candidates_{}".format(buildername, partial_version, partial_info["buildNumber"], chunk) %}
 -
     taskId: "{{ stableSlugId(partial_beetmover_buildername) }}"
     requires:
