--- releasetasks/test/test_beetmover_l10n_candidates.py
@@ -1,133 +1,0 @@
-import unittest
-
-from releasetasks.test import make_task_graph, PVT_KEY_FILE, \
-    do_common_assertions, get_task_by_name
-from release.platforms import buildbot2ftp
-
-
-class TestBeetmover110nCandidates(unittest.TestCase):
-    maxDiff = 30000
-    graph = None
-    # we will end up with one task for each platform
-    tasks = None
-    en_US_config = {
-        "platforms": {
-            "macosx64": {"task_id": "xyz"},
-            "win32": {"task_id": "xyy"}
-        }
-    }
-    l10n_config = {
-        "platforms": {
-            "win32": {
-                "en_us_binary_url": "https://queue.taskcluster.net/something/firefox.exe",
-                "locales": ["de", "en-GB", "zh-TW"],
-                "chunks": 1,
-            },
-            "macosx64": {
-                "en_us_binary_url": "https://queue.taskcluster.net/something/firefox.tar.xz",
-                "locales": ["de", "en-GB", "zh-TW"],
-                "chunks": 1,
-            },
-
-        },
-        "changesets": {
-            "de": "default",
-            "en-GB": "default",
-            "zh-TW": "default",
-        },
-    }
-
-    def setUp(self):
-        self.graph = make_task_graph(
-            version="42.0b2",
-            next_version="42.0b3",
-            appVersion="42.0",
-            buildNumber=3,
-            source_enabled=False,
-            updates_enabled=False,
-            bouncer_enabled=False,
-            push_to_candidates_enabled=True,
-            postrelease_version_bump_enabled=False,
-            en_US_config=self.en_US_config,
-            l10n_config=self.l10n_config,
-            partial_updates={
-                "38.0": {
-                    "buildNumber": 1,
-                },
-                "37.0": {
-                    "buildNumber": 2,
-                },
-            },
-            balrog_api_root="https://fake.balrog/api",
-            signing_class="release-signing",
-            branch="mozilla-beta",
-            product="firefox",
-            repo_path="releases/mozilla-beta",
-            revision="abcdef123456",
-            release_channels=["beta"],
-            signing_pvt_key=PVT_KEY_FILE,
-        )
-        self.tasks = {
-            'win32': get_task_by_name(
-                self.graph, "release-{}_{}_{}_l10n_repack_beetmover_candidates_1".format("mozilla-beta",
-                                                                                         "firefox",
-                                                                                         'win32')
-            ),
-            'macosx64': get_task_by_name(
-                self.graph, "release-{}_{}_{}_l10n_repack_beetmover_candidates_1".format("mozilla-beta",
-                                                                                         "firefox",
-                                                                                         'macosx64')
-            ),
-        }
-
-    def test_common_assertions(self):
-        do_common_assertions(self.graph)
-
-    def test_provisioner(self):
-        for platform, task in self.tasks.iteritems():
-            self.assertEqual(task["task"]["provisionerId"], "aws-provisioner-v1")
-
-    def test_worker_type(self):
-        for platform, task in self.tasks.iteritems():
-            self.assertEqual(task["task"]["workerType"], "opt-linux64")
-
-    def test_scopes_present(self):
-        for platform, task in self.tasks.iteritems():
-            self.assertFalse("scopes" in task)
-
-    def test_platform_in_command(self):
-        for platform, task in self.tasks.iteritems():
-            command = task['task']['payload']['command']
-            self.assertTrue("--platform {}".format(buildbot2ftp(platform)) in "".join(command))
-
-    def test_version_in_command(self):
-        for platform, task in self.tasks.iteritems():
-            command = task['task']['payload']['command']
-            self.assertTrue("--version 42.0b2" in "".join(command))
-
-    def test_app_version_in_command(self):
-        for platform, task in self.tasks.iteritems():
-            command = task['task']['payload']['command']
-            self.assertTrue("--app-version 42.0" in "".join(command))
-
-    def test_locale_in_command(self):
-        for platform, task in self.tasks.iteritems():
-            command = task['task']['payload']['command']
-            self.assertTrue("--locale de --locale en-GB --locale zh-TW" in "".join(command))
-
-    def test_taskid_in_command(self):
-        for platform, task in self.tasks.iteritems():
-            l10n_artifact_task = get_task_by_name(
-                self.graph,
-                "release-mozilla-beta_firefox_{}_l10n_repack_artifacts_1".format(platform)
-            )
-            command = task['task']['payload']['command']
-            self.assertTrue("--taskid {}".format(l10n_artifact_task['taskId']) in "".join(command))
-
-    def test_graph_scopes(self):
-        expected_graph_scopes = set([
-            "queue:task-priority:high",
-            "queue:define-task:aws-provisioner-v1/opt-linux64",
-            "queue:create-task:aws-provisioner-v1/opt-linux64"
-        ])
-        self.assertTrue(expected_graph_scopes.issubset(self.graph["scopes"]))
